+++ app/interactives/inflation-impact-calculator/page.tsx
@@ -7,7 +7,7 @@ import { Label } from "@/app/ui/components/label"
 import { CustomSlider } from "@/app/ui/components/slider"
 import { Badge } from "@/app/ui/components/badge"
 import { BiSolidUpArrow, BiSolidDownArrow } from "react-icons/bi";
-import ThemeToggle from "@/app/ui/theme-toggle"; 
+import ThemeToggle from "@/app/lib/theme-toggle";
 
 export default function InflationCalculator() {
   const [initialPrice, setInitialPrice] = useState(100)
